requirements.txt
Django==4.2.10
gunicorn==21.2.0
whitenoise==6.6.0
psycopg2-binary==2.9.9
dj-database-url==2.1.0
python-dotenv==1.0.0
pillow==10.4.0
cloudinary==1.38.0
django-cloudinary-storage==0.3.0